+++ src/main/java/io/hhplus/tdd/point/entity/PointHistory.java
@@ -1,4 +1,6 @@
-package io.hhplus.tdd.point;
+package io.hhplus.tdd.point.entity;
+
+import io.hhplus.tdd.point.model.TransactionType;
 
 public record PointHistory(
         long id,
